feat/stats: stash in server component
